requirements.txt
Django
Pillow
django-jazzmin
django-ckeditor
django-colorfield
